Refactor nginx package
- Rename the nginx package to configs
- Extract the nginx package from the configs package
- Export less in the nginx package
