+++ internal/configs/config.go
@@ -1,4 +1,4 @@
-package nginx
+package configs
 
 import (
 	"strings"
