+++ internal/configs/configurator_test.go
@@ -1,4 +1,4 @@
-package nginx
+package configs
 
 import (
 	"net/http"
@@ -6,7 +6,7 @@ import (
 	"sort"
 	"testing"
 
-	"github.com/nginxinc/kubernetes-ingress/internal/nginx/plus"
+	"github.com/nginxinc/kubernetes-ingress/internal/nginx"
 	api_v1 "k8s.io/api/core/v1"
 	extensions "k8s.io/api/extensions/v1beta1"
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -528,8 +528,8 @@ func createTestConfigurator() (*Configurator, error) {
 	if err != nil {
 		return nil, err
 	}
-	ngxc := NewNginxController("/etc/nginx", "nginx", true)
-	apiCtrl, err := plus.NewNginxAPIController(&http.Client{}, "", true)
+	ngxc := nginx.NewNginxController("/etc/nginx", "nginx", true)
+	apiCtrl, err := nginx.NewNginxAPIController(&http.Client{}, "", true)
 	if err != nil {
 		return nil, err
 	}
@@ -545,8 +545,8 @@ func createTestConfiguratorInvalidIngressTemplate() (*Configurator, error) {
 	if err := templateExecutor.UpdateIngressTemplate(&invalidIngressTemplate); err != nil {
 		return nil, err
 	}
-	ngxc := NewNginxController("/etc/nginx", "nginx", true)
-	apiCtrl, _ := plus.NewNginxAPIController(&http.Client{}, "", true)
+	ngxc := nginx.NewNginxController("/etc/nginx", "nginx", true)
+	apiCtrl, _ := nginx.NewNginxAPIController(&http.Client{}, "", true)
 	return NewConfigurator(ngxc, NewDefaultConfig(), apiCtrl, templateExecutor, false), nil
 }
