+++ internal/configs/convert_test.go
@@ -1,4 +1,4 @@
-package nginx
+package configs
 
 import (
 	"reflect"
