+++ internal/configs/templates/nginx-plus.tmpl
