+++ internal/configs/templates/nginx.tmpl
