+++ internal/nginx/verify.go
@@ -1,8 +1,10 @@
-package verify
+package nginx
 
 import (
+	"bytes"
 	"context"
 	"fmt"
+	"html/template"
 	"io/ioutil"
 	"net"
 	"net/http"
@@ -12,15 +14,15 @@ import (
 	"github.com/golang/glog"
 )
 
-// Client is a client for verifying the config version.
-type Client struct {
+// verifyClient is a client for verifying the config version.
+type verifyClient struct {
 	client     *http.Client
 	maxRetries int
 }
 
-// NewClient returns a new client pointed at the config version socket.
-func NewClient() *Client {
-	return &Client{
+// newVerifyClient returns a new client pointed at the config version socket.
+func newVerifyClient() *verifyClient {
+	return &verifyClient{
 		client: &http.Client{
 			Transport: &http.Transport{
 				DialContext: func(_ context.Context, _, _ string) (net.Conn, error) {
@@ -34,7 +36,7 @@ func NewClient() *Client {
 
 // GetConfigVersion get version number that we put in the nginx config to verify that we're using
 // the correct config.
-func (c *Client) GetConfigVersion() (int, error) {
+func (c *verifyClient) GetConfigVersion() (int, error) {
 	resp, err := c.client.Get("http://config-version/configVersion")
 	if err != nil {
 		return 0, fmt.Errorf("error getting client: %v", err)
@@ -58,7 +60,7 @@ func (c *Client) GetConfigVersion() (int, error) {
 
 // WaitForCorrectVersion calls the config version endpoint until it gets the expectedVersion,
 // which ensures that a new worker process has been started for that config version.
-func (c *Client) WaitForCorrectVersion(expectedVersion int) error {
+func (c *verifyClient) WaitForCorrectVersion(expectedVersion int) error {
 	sleep := 25 * time.Millisecond
 	for i := 1; i <= c.maxRetries; i++ {
 		time.Sleep(sleep)
@@ -74,4 +76,49 @@ func (c *Client) WaitForCorrectVersion(expectedVersion int) error {
 		}
 	}
 	return fmt.Errorf("could not get expected version: %v", expectedVersion)
+}
+
+const configVersionTemplateString = `server {
+    listen unix:/var/run/nginx-config-version.sock;
+    access_log off;
+
+    location /configVersion {
+        return 200 {{.ConfigVersion}};
+    }
+}
+map $http_x_expected_config_version $config_version_mismatch {
+	"{{.ConfigVersion}}" "";
+	default "mismatch";
+}`
+
+// verifyConfigGenerator handles generating and writing the config version file.
+type verifyConfigGenerator struct {
+	configVersionTemplate *template.Template
+}
+
+// newVerifyConfigGenerator builds a new ConfigWriter - primarily parsing the config version template.
+func newVerifyConfigGenerator() (*verifyConfigGenerator, error) {
+	configVersionTemplate, err := template.New("configVersionTemplate").Parse(configVersionTemplateString)
+	if err != nil {
+		return nil, err
+	}
+	return &verifyConfigGenerator{
+		configVersionTemplate: configVersionTemplate,
+	}, nil
+}
+
+// GenerateVersionConfig generates the config version file.
+func (c *verifyConfigGenerator) GenerateVersionConfig(configVersion int) ([]byte, error) {
+	var configBuffer bytes.Buffer
+	templateValues := struct {
+		ConfigVersion int
+	}{
+		configVersion,
+	}
+	err := c.configVersionTemplate.Execute(&configBuffer, templateValues)
+	if err != nil {
+		return nil, err
+	}
+
+	return configBuffer.Bytes(), nil
 }
